requirements.txt
aiogram>=3.15.0
httpx>=0.28.0
supabase>=2.11.0
pydantic>=2.4.1
pydantic-settings>=2.7.0
pyzbar>=0.1.9
Pillow>=11.0.0
python-dotenv>=1.0.0
